Fix tiff file loading
The issue we found was the following: In the TIFF file the image data
is stored in strips (a strip can represent the whole image, a single
scanline, or multiple scanlines). In our case each strip represented
a scanline. The TIFF file contains a table with the file offset of each
strip, and a separate table with the byte count in each strip.
According to the specification, both table can contain either 16bit
or 32bit integer values. However tiny_dng_loader assumed they are
always 32bit values. In our case the TIFF file had 32bit offsets (which
was parsed correctly) but contained 16bit byte counts. The code parsed
that raw data as 32 bit values, so the data was garbage (the values
became HUGE, which caused huge virtual memory allocations, slowing
down the loading) and in the middle of the image, it run out of data
and started reading data from the next table (which happened to be the
strip offsets). The very first value in the strip offset table was 8,
which was interpreted as the byte count in the center strip/scanline,
hence producing the horisontal line artifact (since the decoder used
only 8 bytes to produce the whole scanline).

To fix the issue, we need to keep track of the type of these two tables
and the load the values accordingly.

Also removed the "strip_byte_count" field... it was not really used
(and also parsed wrongly anyway... probably it was some leftover).